Core vadim openapi (#60)
* OpenApi

* OpenAPI validate register player

* OpenAPI register player

* OpenAPI register player

+++ core/internal/handlers/register_player_in_sport_section.go
@@ -8,7 +8,7 @@ import (
 
 type (
 	RegisterPlayerInSportSectionService interface {
-		RegisterPlayerInSportSection(ctx context.Context, id int64) (string, error)
+		RegisterPlayerInSportSection(ctx context.Context, id int) (string, error)
 	}
 
 	RegisterPlayerInSportSectionHandler struct {
